% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_as_tex.R
\name{addlinespace}
\alias{addlinespace}
\title{Add addlinespace commands to latex table}
\usage{
addlinespace(table, terms)
}
\arguments{
\item{table}{Latex table}

\item{terms}{Strings, which terms to add addlinespace after}
}
\description{
Default output of latex tables will add some addlinespace commands for you,
but you can provide which terms you want to add the spaces after with this
command. Note that special characters like curly braces should be escaped
with two backslashes.
}
