% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_helpers.R
\name{as_prior_list}
\alias{as_prior_list}
\title{Title}
\usage{
as_prior_list(prior_specs)
}
\arguments{
\item{prior_specs}{brmsprior object}
}
\value{
List of brmsprior objects
}
\description{
Title
}
