% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward_difference_code.R
\name{backward_difference_code}
\alias{backward_difference_code}
\title{Backward difference code}
\usage{
backward_difference_code(n_levels)
}
\arguments{
\item{n_levels}{Number of factor levels}
}
\value{
Backward difference contrast matrix
}
\description{
Compares the mean of level k to level k-1. Differs from reverse helmert coding
because it doesn't compare k to mean(k_0, ..., k_i-1). Mathematically just
flipping the signs of the matrix from forward difference coding.
}
\details{
Example interpretation for a 4 level factor:
Intercept = Grand mean (mean of the means of each level)
grp1 = mean(grp2) - mean(grp1)
grp2 = mean(grp3) - mean(grp2)
grp3 = mean(grp4) - mean(grp3)
}
\examples{
mydf <- data.frame(
   grp = factor(c(rep("F1",5),rep("F2",5),rep("F3",5),rep("F4",5))),
   resp = c(seq(1,5), seq(5,9), seq(10,14), seq(15,19))
)

mydf \%>\%
  dplyr::group_by(grp) \%>\%
  dplyr::summarize(mu = mean(resp)) \%>\%
  dplyr::ungroup() \%>\%
  dplyr::mutate(grand_mean = mean(mu))

summary(lm(resp ~ grp,
           data = mydf,
           contrasts = enlist_contrasts(mydf, grp ~ backward_difference_code)))
}
