% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/png_watermarking.R
\name{binary_to_string}
\alias{binary_to_string}
\title{Convert binary character vector to string}
\usage{
binary_to_string(binary_vector)
}
\arguments{
\item{binary_vector}{Character vector of binary encoded values}
}
\value{
String conversion of binary vector
}
\description{
Given a character vector of 8 bit binary representations, convert to a string
}
