% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brms_fit_check.R
\name{brmsfit_needs_refit2}
\alias{brmsfit_needs_refit2}
\title{Diagnostics for model refitting}
\usage{
brmsfit_needs_refit2(
  fit,
  sdata = NULL,
  scode = NULL,
  data = NULL,
  algorithm = NULL,
  verbose = FALSE,
  formula = NULL,
  backend = NULL,
  prior = NULL,
  control = NULL
)
}
\arguments{
\item{fit}{Previously fit model}

\item{sdata}{Generated standata}

\item{scode}{Generated stancode}

\item{data}{Validated data}

\item{algorithm}{algorithm choice}

\item{verbose}{verbosity,default TRUE}

\item{formula}{New formula}

\item{backend}{backend choice}

\item{prior}{List of priors}

\item{control}{brm control argument}
}
\value{
List of logical values denoting which diagnostics indicate the model
should be refit
}
\description{
Extension of `brms::brmsfit_needs_refit`
}
\details{
Checks for the following things and returns a list with the names below:

- `software_version_changed`: Did the brms, cmdstan, cmdstanr, rstan, or
   StanHeaders package versions change compared to the cached model?
- `data_dim_changed`: Did the dimensions (`[rows,columns]`) of the data change?
- `backend_changed`: Did the backend change (eg from cmdstanr to rstan)?
- `formula_changed`: Did the formula change?
- `scode_changed`: Did the generated stan code change?
- `factors_changed`: Did the levels of any factors change?
- `contrasts_changed`: Did the contrasts of any factors change? Variables
that are implicitly coerced to factors (such as character values) are assumed
to use contr.treatment by default.
- `algorithm_changed`: Did the algorithm change?
- `priors_changed`: Did any priors change?
}
