% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloglog_estimates.R
\name{cloglog_estimates}
\alias{cloglog_estimates}
\title{Sample cloglog estimates}
\usage{
cloglog_estimates(table, reference = 1)
}
\arguments{
\item{table}{Contingency table from `xtabs`}

\item{reference}{Which row should be used as the reference level, defaults to
1}
}
\value{
Data frame with cloglog thresholds and effect estimates.
}
\description{
Probit analog to `sostools::odds_ratios`. Given a 2d contingency table,
calculate the cloglog estimates between a reference level (row) and the
other rows. Does not currently support confidence intervals.
}
