% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brms_fit_check.R
\name{compare_brmsfit_formulas}
\alias{compare_brmsfit_formulas}
\title{Check brmsfit formulas}
\usage{
compare_brmsfit_formulas(f1, f2)
}
\arguments{
\item{f1}{New formula}

\item{f2}{Cached formula from `fit[["formula"]]`}
}
\description{
Goes through each component of the formulas used in brmsfit objects
and tries to determine which things are different.
}
\details{
Will report the following:

 - Whether the main part of the formula is different
 - Whether the pforms (additional parameters like sigma or disc) are
 different. Specifically, whether and which pforms were added, removed, or
 modified.
 - Whether the response variable changed
 - Whether the likelihood family changed
 - Whether the mecor changed (I dont know what this is though)
}
