% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brms_fit_check.R
\name{compare_dv}
\alias{compare_dv}
\title{Compare response variables}
\usage{
compare_dv(fit, data, verbose = FALSE)
}
\arguments{
\item{fit}{Model fit}

\item{data}{Dataframe}

\item{verbose}{Whether to print differences in attributes, if any}
}
\value{
Logical value depending on whether the values of the response
variable differ
}
\description{
Compare response variables
}
