% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concordancy_pairs.R
\name{concordancy_pairs}
\alias{concordancy_pairs}
\title{Calculate concordancy pairs}
\usage{
concordancy_pairs(contingency_table)
}
\arguments{
\item{contingency_table}{Contingency table of counts}
}
\value{
Named vector of number of Concordant and Discordant pairs.
}
\description{
Calcualtes the number of concordant and number of discordant pairs in
a 2D contingency table. Ascending or descending order does not matter for
this calculation. Modeled after Agresti 2010 pg 186. Generally, if there
are more concordant pairs than discordant pairs (C - D > 0), then the
association of the ordinal variables is positive. If more discordant
(C - D < 0), then the association is negative.
}
