% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast_code.R
\name{contrast_code}
\alias{contrast_code}
\title{Contrast code factors}
\usage{
contrast_code(factor_col, code_by = NA, ...)
}
\arguments{
\item{factor_col}{The factor column to use, eg data$gender}

\item{code_by}{Either a matrix or a function}

\item{...}{Additional arguments to be passed to functional_code, specifically,
which level you want the reference level to be}
}
\value{
A contrast coding matrix with labels and proper reference level
}
\description{
Helper to do contrast coding. There are two options:
 - Manually specify a matrix for code_by (implements manual_code). Reference level
 is automatically set to the row that's always negative.
 - Specify a style of contrast coding as a function (implements functional_code).
 Label of the reference level should be specified in ...
}
