% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrasts_as_latex.R
\name{contrasts_as_latex}
\alias{contrasts_as_latex}
\title{Make latex table of contrasts}
\usage{
contrasts_as_latex(factor_col)
}
\arguments{
\item{factor_col}{A factor column, eg data$gender}
}
\value{
A latex table string
}
\description{
Given a factor column, print the contrast matrices used
}
