% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord_aspect.R
\name{coord_aspect}
\alias{coord_aspect}
\title{Fixed aspect ratio relative to X/Y limits}
\usage{
coord_aspect(ratio = 1, xlim = NULL, ylim = NULL, expand = FALSE, clip = "on")
}
\arguments{
\item{ratio}{Numeric ratio, default 1 (for square). Ratios should be of the
form height / width.}

\item{xlim}{Numeric vector of length 2, limits for x axis}

\item{ylim}{Numeric vector of length 2, limits for y axis}

\item{expand}{Logical, default FALSE, whether to expand}

\item{clip}{Whether to clip to plot, default "on"}
}
\value{
coordinate ggproto
}
\description{
Uses a given aspect ratio for a ggplot based on the limits provided.
}
\details{
If the range of X values is 10 to 200 and the range of Y values is 20 to 60,
then a square aspect ratio would require a ratio of (200 - 10) / (60 - 40).
A landscape 16:9 aspect ratio would be (9/16) * (200-10) / (60-40). This is
an annoying amount of things to type out if you're fiddling around with the
limits and aspect ratio.
}
