% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloglog.R
\name{dcloglog}
\alias{dcloglog}
\title{PDF for cloglog distribution}
\usage{
dcloglog(x, location = 0, scale = 1)
}
\arguments{
\item{x}{vector of quantiles}

\item{location}{location parameter}

\item{scale}{scale parameter}
}
\value{
Probability density function
}
\description{
Works like `dlogis`, returns the probability density at a given value.
Value is equal to the derivative of the CDF defined by
}
