% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompose_contrasts.R
\name{decompose_contrasts}
\alias{decompose_contrasts}
\title{Decompose contrasts into new columns}
\usage{
decompose_contrasts(
  model_data,
  factor_col,
  extract.seq = NULL,
  extract.to = NULL
)
}
\arguments{
\item{model_data}{Data frame given to model}

\item{factor_col}{Which column (as a string)}

\item{extract.seq}{integer vector of which components to get, use NULL (default)
to extract all of them}

\item{extract.to}{Optional names to give the new columns, must equal number of
extracted components.}
}
\value{
model_data but with new columns corresponding to the numeric coding
of the given factor's contrasts
}
\description{
First draft of pulling out the contrast coding numbers from the model matrix
and adding them into new columns. Currently only handles 1 column at a time
and does not work within `mutate`, which is where it would be most helpful.
}
