% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_propodds.R
\name{.add_ordinal_result}
\alias{.add_ordinal_result}
\title{Merge CLMM object}
\usage{
.add_ordinal_result(clm_obj, results)
}
\arguments{
\item{clm_obj}{CLM or CLMM object from the ordinal package}

\item{results}{Results from .run_single_glm}
}
\value{
Extended dataframe
}
\description{
Adds the confidence intervals for the CLM estimated effects to compare the
individual binary regressions against. Note that interactions must be
specified in the same order in both model equations. Will check if the
confidence intervals from the binary regressions include the confidence
region from the CLM/CLMM estimates.
}
