% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_if_valid_contrmat.R
\name{.check_if_valid_contrmat}
\alias{.check_if_valid_contrmat}
\title{Check if valid contrast}
\usage{
.check_if_valid_contrmat(contr_mat)
}
\arguments{
\item{contr_mat}{Contrast matrix}
}
\description{
Sometimes a user might pass a matrix that can't actually be used, in which
case we can avoid some calculations by stopping earlier.
}
