% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_formula.R
\name{.check_if_valid_formula}
\alias{.check_if_valid_formula}
\title{Formula validator}
\usage{
.check_if_valid_formula(formula, char_formula, no_matrix_string)
}
\arguments{
\item{formula}{Raw formula}

\item{char_formula}{Character string of the formula from deparse}

\item{no_matrix_string}{String without any matrix calls}
}
\value{
Nothing, throws an error if any are found
}
\description{
Given a formula (and various preprocessed versions of it), run diagnostics
to ensure that it's a valid formula.
}
