% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enlist_contrasts.R
\name{.check_remaining_factors}
\alias{.check_remaining_factors}
\title{Check for unspecified factors}
\usage{
.check_remaining_factors(model_data, specified_vars)
}
\arguments{
\item{model_data}{Model data}

\item{specified_vars}{variables specified by the user from formulas}
}
\value{
nothing, just sends a message if needed
}
\description{
Sends a message if the user has factor columns in their model data frame
that weren't specified with the others.
}
