% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glimpse_contrasts.R
\name{.cols_where}
\alias{.cols_where}
\title{Get columns where}
\usage{
.cols_where(model_data, fx, use.names = FALSE, return.names = FALSE)
}
\arguments{
\item{model_data}{Model data}

\item{fx}{Function to apply, must be something that returns a logical value.
Usually either `is.factor` or `is.ordered`}

\item{use.names}{Whether the resulting vector should be named}

\item{return.names}{Whether names (where the fx returns TRUE) should be
returned instead of a logical vector. Overwrites use.names.}
}
\value{
optionally named logical vector or character vector
}
\description{
Helper to avoid the use of tidyselect and dplyr::select, returns either a
logical vector (optionally named) or a character vector of which columns
satisfy the given function
}
