% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_coefficient.R
\name{.format_brms_coef}
\alias{.format_brms_coef}
\title{Title}
\usage{
.format_brms_coef(coeff, clist)
}
\arguments{
\item{coeff}{Which coefficient to format}

\item{clist}{The list of coefficient data to use}
}
\value{
Formatted coefficient string
}
\description{
Title
}
