% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_to_char.R
\name{.formula_to_char}
\alias{.formula_to_char}
\title{formula to character error handling}
\usage{
.formula_to_char(formulas)
}
\arguments{
\item{formulas}{Series of formulas passed by `enlist_contrasts`}
}
\value{
list of formulas coerced to character, or an error
}
\description{
Adds a more helpful hint when handling the initial coercion in `enlist_contrasts`
}
