% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odds_ratios.R
\name{.generate_or_table}
\alias{.generate_or_table}
\title{Generate odds ratio table}
\usage{
.generate_or_table(r, c, odds_ratio_matrix, standard_error_matrix, bonf)
}
\arguments{
\item{r}{Number of rows}

\item{c}{Number of columns}

\item{odds_ratio_matrix}{Odds ratios}

\item{standard_error_matrix}{Standard errors}

\item{bonf}{Logical, whether bonferroni correction should be performed}
}
\description{
Given an odds ratio matrix and its associated standard errors, calculate
confidence intervals and report in a table like Agresti 2010 pg 29
(Table 2.6)
}
