% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enlist_contrasts.R
\name{.get_if_exists}
\alias{.get_if_exists}
\title{Get value if it exists}
\usage{
.get_if_exists(scheme_argument, var_envir)
}
\arguments{
\item{scheme_argument}{Argument used in the scheme}

\item{var_envir}{Environment to look up value}
}
\value{
Value of a variable, or the original string cleaned up
}
\description{
Given a name and an environment, lookup its value. If it isn't a variable name
it will do some string preprocessing before returning the value. Later
functions use mostly string representations anyways.
}
