% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_parameters.R
\name{.make_parameters}
\alias{.make_parameters}
\title{Make parameters for contrast code call}
\usage{
.make_parameters(
  formula,
  params = list(factor_col = NA, code_by = NA, reference_level = NA, intercept_level =
    NA, drop_trends = NA),
  EMBEDDED = FALSE
)
}
\arguments{
\item{formula}{Formula given by user}

\item{params}{Parameter accumulator}

\item{EMBEDDED}{Whether the current operation is embedded in another, helps
to handle operator precedence with *}
}
\value{
Named list of parameters that can be evaluated in `enlist_contrasts`
}
\description{
Given a formula, recursively go through the abstract syntax tree and
extract the necessary parameters for `contrast_code`. While this method is
more involved than extracting from a parsed string representation, I think
it handles matrix calls better since it plucks the call right out of
the syntax tree.
}
