% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glimpse_contrasts.R
\name{.make_placeholder_glimpse}
\alias{.make_placeholder_glimpse}
\title{Make glimpse for 1-level factors}
\usage{
.make_placeholder_glimpse(model_data, one_level_factors)
}
\arguments{
\item{model_data}{Model data}

\item{one_level_factors}{Which factors are one level}
}
\value{
A tibble with limited information about one level factors
}
\description{
If the user wants factors with only one level included, this will create
the table to present that information. It's a lot of NAs because contrasts
aren't defined for only 1 level.
}
