% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_coercion_messages.R
\name{.msg_if_coerced_to_factors}
\alias{.msg_if_coerced_to_factors}
\title{Alert user when factor coercion happens}
\usage{
.msg_if_coerced_to_factors(which_to_factors)
}
\arguments{
\item{which_to_factors}{Names of columns that have been coerced to factors}
}
\value{
Nothing, sends a message if needed
}
\description{
Sends a message if columns in the dataset have been coerced to a factor,
useful if you accidentally set a column with continuous data to a factor.
}
