% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_coercion_messages.R
\name{.msg_if_ordered_reset}
\alias{.msg_if_ordered_reset}
\title{Alert user when setting ordered factors}
\usage{
.msg_if_ordered_reset(model_data, vars_in_model)
}
\arguments{
\item{model_data}{Data frame to be used with the model}

\item{vars_in_model}{Variables to check against}
}
\value{
Nothing, sends a message
}
\description{
Ordered factors use orthogonal polynomials (contr.poly) by default with
k number of levels -1 (k-1) of trend comparisons. If you use these functions
to set the contrasts to something else, the ordered class (hence ordering
to the levels) will remain but the contrasts will be set to something else.
}
