% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_drop_sequence.R
\name{.parse_drop_sequence}
\alias{.parse_drop_sequence}
\title{Parse comparisons to drop}
\usage{
.parse_drop_sequence(scheme_argument, var_envir)
}
\arguments{
\item{scheme_argument}{A scheme argument of the form "a:b", where a and b
can correspond to variable names bound in var_envir or numbers.}

\item{var_envir}{Environment to look up values if a and b are variable names}
}
\value{
A sequence of numbers
}
\description{
This parses any sequences following a `-` operator and expands them to
a numeric vector.
}
