% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_contrasts.R
\name{.reinstate_dropped_trends}
\alias{.reinstate_dropped_trends}
\title{Ignore attempt to drop trends in set_contrasts}
\usage{
.reinstate_dropped_trends(formulas)
}
\arguments{
\item{formulas}{formulas passed to `set_contrasts`}
}
\value{
Cleaned up formulas as needed
}
\description{
Will send a warning if it detects dropped trends with orthogonal polynomial
contrasts. If you use a different coding scheme it just gets ignored anyways.
}
