% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_propodds.R
\name{.run_single_glm}
\alias{.run_single_glm}
\title{Run binary regression}
\usage{
.run_single_glm(
  ord_data,
  mdl_fx,
  coef_fx,
  model_formula,
  response_var,
  j,
  .link = "logit",
  control = list()
)
}
\arguments{
\item{ord_data}{Ordinal dataframe}

\item{mdl_fx}{Model function to use, either glm or glmer}

\item{coef_fx}{Coefficient function to use, either fixef or coef}

\item{model_formula}{Model formula to evaluate}

\item{response_var}{LHS of the formula}

\item{j}{Cutpoint value}

\item{.link}{Link function, default is logit (ie logistic regression)}

\item{control}{Control to use for mixed models, will default to bobyqa if not specified}
}
\value{
Dataframe with binary regression results
}
\description{
Split data at cutpoint j and run a binary regression with the given
link function.
}
