% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enlist_contrasts.R
\name{.scrub_scheme}
\alias{.scrub_scheme}
\title{Remove whitespace}
\usage{
.scrub_scheme(coding_scheme)
}
\arguments{
\item{coding_scheme}{coding scheme in enlist_contrasts}
}
\value{
new coding scheme char vector trimmed whitespace
}
\description{
small helper to trim whitespace
}
