% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_as_tex_brmsfit.R
\name{.summary_as_tex.brmsfit}
\alias{.summary_as_tex.brmsfit}
\title{BRMS Table output}
\usage{
.summary_as_tex.brmsfit(
  model,
  caption = NA,
  label = NA,
  include_prior = FALSE,
  table.envir = "table",
  tabular.envir = "tabular",
  ci_level = 0.95,
  use_ROPE = FALSE
)
}
\arguments{
\item{model}{brmsfit object}

\item{caption}{Caption for LaTeX table}

\item{label}{Label for LaTeX cross referencing}

\item{include_prior}{Logical, should priors be included in table?}

\item{table.envir}{Table environment to use}

\item{tabular.envir}{Tabular environment to use}

\item{ci_level}{Defaults to .95}

\item{use_ROPE}{ROPE object from bayestestR::rope, or TRUE to run this with the
default settings. Adds the % in ROPE to table after confidence intervals. FALSE
to not include.}
}
\value{
Formatted table
}
\description{
BRMS Table output
}
