% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switch_reference_level.R
\name{.switch_reference_level}
\alias{.switch_reference_level}
\title{Switch reference level in contrast matrix}
\usage{
.switch_reference_level(contrast_matrix, n_levels, reference)
}
\arguments{
\item{contrast_matrix}{Contrast matrix, unlabeled}

\item{n_levels}{Number of levels in the factor}

\item{reference}{Which row to use as the reference level}
}
\value{
A matrix with the correct values for the reference level
}
\description{
Switch rows to set the reference level properly
}
