% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_helpers.R
\name{drop_duplicated_priors}
\alias{drop_duplicated_priors}
\title{Drop duplicated prior specifications}
\usage{
drop_duplicated_priors(priors)
}
\arguments{
\item{priors}{Object of type `brmsprior`}
}
\value{
brmsprior without duplicates. Note that the first instance in the
brmsprior dataframe is retained.
}
\description{
brm will throw an error if multiple priors are explicitly set for the same
parameter. This helper uses `brms:::duplicated.brmsprior` to filter out any
duplicates.
}
