% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enlist_contrasts.R
\name{enlist_contrasts}
\alias{enlist_contrasts}
\title{List of contrast matrices}
\usage{
enlist_contrasts(model_data, ..., verbose = TRUE)
}
\arguments{
\item{model_data}{Data frame you intend on passing to your model}

\item{...}{A series of 2 sided formulas with factor name on the LHS and
desired contrast scheme on the RHS, reference levels can be set with + and the
intercept can be overwritten with * (+ should come first if both are set)}

\item{verbose}{Should messages be sent to user? Defaults to TRUE, MUST be
a named argument, or it will get parsed as part of ...}
}
\value{
List of named contrast matrices
}
\description{
Returns a list of contrast matrices to use with modeling functions directly.
}
\details{
Typically model functions like lm will have a contrasts argument where you can
set the contrasts at model run time, rather than having to manually change
the contrasts on the underlying factor columns in your data. If you prefer
this way, you can use this to generate the named list of contrast matrices.
}
\examples{
my_df <- mtcars
my_df$gear = factor(my_df$gear)
my_df$carb = factor(my_df$carb)

# Use formulas where left hand side is the factor column name
# and the right hand side is the contrast scheme you want to use
enlist_contrasts(my_df,
    gear ~ scaled_sum_code, # Using helpers from this package
    carb ~ helmert_code)

# Add reference levels with +
enlist_contrasts(my_df,
    gear ~ scaled_sum_code + 5,
    carb ~ contr.sum + 6)
# Manually specifying matrix also works
enlist_contrasts(my_df,
    gear ~ matrix(c(1,-1,0,0,-1,1), nrow = 3),
    carb ~ forward_difference_code)

# User matrices can be assigned to a variable first, but this may make the
# comparison labels confusing. You should rename them manually to something
# that makes sense. This will invoke manual_code, so reference levels
# specified with + will be ignored.
my_gear_contrasts <- matrix(c(1,-1,0,0,-1,1), nrow = 3)
colnames(my_gear_contrasts) <- c("CMP1", "CMP2")
enlist_contrasts(my_df,
    gear ~ my_gear_contrasts,
    carb ~ forward_difference_code)


# Will inform you if there are factors you didn't set
enlist_contrasts(my_df, gear ~ scaled_sum_code)

}
