% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_bam.R
\name{fit_bam}
\alias{fit_bam}
\title{Fit GAMM with diagnostic checks}
\usage{
fit_bam(mdl_formula, file, data, ...)
}
\arguments{
\item{mdl_formula}{Formula passed to `bam`}

\item{file}{Path to a file to save the model to}

\item{data}{Dataframe to use}

\item{...}{Additional arguments passed to `bam` (method, nthreads, etc.)}
}
\value{
the result of `bam`
}
\description{
Runs `mgcv::bam` but also saves the result to a file and runs diagnostics
to check if the model specification or data has changed since the last time
it was fit.
}
\examples{
\dontrun{
gam_mdl <- fit_bam(
y ~ grp +
     s(t, grp, bs = "fs", k = 15) +
     s(t, grp_2, bs = "fs", by = grp, k = 15),
"Models/gam_mdl.rds",
data = mdl_data,
discrete = TRUE,
method = "fREML",
nthreads = 8
)

}
}
