% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brms_fit_check.R
\name{fit_kfold}
\alias{fit_kfold}
\title{Wrapper to run kfold cross validation with preset fold assignments}
\usage{
fit_kfold(
  mdl,
  kfold_path,
  fold_assignments,
  k = 5,
  nworkers = 5,
  cores_per_fold = 1,
  nchains = 4
)
}
\arguments{
\item{mdl}{`brmsfit` object}

\item{kfold_path}{Path to save the kfold-cv model to. Note that this is
going to be a list of brmsfit objects with some overhead functionality on top.}

\item{fold_assignments}{Integer vector of length `nrow(mdl$data)` assigning
each row/observation to a fold. This is done outside of this function to
ensure that the folds are consistent and to allow flexibility in however the
user decides to stratify the folds. Max value must equal `k`}

\item{k}{Number of folds}

\item{nworkers}{Number of workers to use for parallelization}

\item{cores_per_fold}{Number of cores to use for each model fitting job}

\item{nchains}{Number of chains to fit for each model. If `k` is very large,
you might want to lower this from 4 to 2.}
}
\value{
`kfold_path`, the PATH to a saved kfold object
}
\description{
Given a `brmsfit` object, run kfold cross validation on that model. Each
row in the dataframe should be associated with an integer assignment provided
by `fold_assignments`.
}
\details{
`k` should be specified to the desired number of folds and should equal the
max number in `fold_assignments`. The model will be refit `k` times, which
can take a very long time. This function will use `future` to parallelize
the models. `nworkers` should be set to the same value as `k` for maximum
efficiency, but can be set to a lower value if there are many folds and chains.

For example, to fit 5 fold models at once, you can have `nworkers=5` and
`cores_per_fold=1`, which will fit each *model* in parallel but each *chain*
of those models sequentially. You can raise `cores_per_fold` to 2 to fit
two chains for each model in parallel. But note that this raises the number
of required *physical* cores from 5 to 10.

Because refitting the model is so expensive, this function will perform a
number of diagnostic tests to verify that the provided `mdl` has changed
compared to the last time kfold CV was run.
}
