% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_coefficient.R
\name{format_coef}
\alias{format_coef}
\title{Format coefficients}
\usage{
format_coef(coeff, clist)
}
\arguments{
\item{coeff}{Which coefficient to format}

\item{clist}{The list of coefficient data to use}
}
\value{
A string of coefficient descriptions
}
\description{
For easily writing coefficient values in text write up. Note this currently
ONLY works for models with zvalues (lm, glm, lmer, glmer). Can be generalized
later but I don't want to right now.
}
