% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_pval.R
\name{format_pval}
\alias{format_pval}
\title{Title
Helper to format p values in inline prose;
src: https://www.eyer.us/post/2019-10-20-f_pvalue/}
\usage{
format_pval(p.value, symbol = "=", use_stars = FALSE)
}
\arguments{
\item{p.value}{double}

\item{symbol}{set to =}

\item{use_stars}{Should add significance stars (.05, .01, .001) default FALSE}
}
\value{
a string for a p value (no stars!)
}
\description{
Title
Helper to format p values in inline prose;
src: https://www.eyer.us/post/2019-10-20-f_pvalue/
}
