% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward_difference_code.R
\name{forward_difference_code}
\alias{forward_difference_code}
\title{Forward difference code}
\usage{
forward_difference_code(n_levels)
}
\arguments{
\item{n_levels}{Number of factor levels}
}
\value{
Forward difference contrast matrix
}
\description{
Compares the mean of level k to level k+1. Differs from helmert coding because
it doesn't compare k to mean(k_i, ..., k_n)
Intercept is the grand mean (mean of cell means)
}
\details{
Example interpretation for a 4 level factor:
\itemize{
\item Intercept = Grand mean (mean of the means of each level)
\item grp1 = mean(grp1) - mean(grp2)
\item grp2 = mean(grp2) - mean(grp3)
\item grp3 = mean(grp3) - mean(grp4)
}
}
\examples{
mydf <- data.frame(
   grp = factor(c(rep("F1",5),rep("F2",5),rep("F3",5),rep("F4",5))),
   resp = c(seq(1,5), seq(5,9), seq(10,14), seq(15,19))
)

mydf \%>\%
  dplyr::group_by(grp) \%>\%
  dplyr::summarize(mu = mean(resp)) \%>\%
  dplyr::ungroup() \%>\%
  dplyr::mutate(grand_mean = mean(mu))

summary(lm(resp ~ grp,
           data = mydf,
           contrasts = enlist_contrasts(mydf, grp ~ scaled_sum_code)))
}
