% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coef.R
\name{get_coef}
\alias{get_coef}
\title{Get coefficient values from list of coefficients}
\usage{
get_coef(coef, clist)
}
\arguments{
\item{coef}{Which coefficient to look up}

\item{clist}{The list of coefficients to use}
}
\value{
Formatted value string for the coefficient
}
\description{
Use get_coeff(coeff, clist) to retrieve the value of coeff from clist
}
\details{
Or use function factor make_coef_getter to make it easier to retrieve values from
specific lists when you have multiple lists.
}
\examples{
mdl <- lm(mpg ~ cyl * gear, data = mtcars) # Create model
coefs <- enlist_coefs(mdl) # Create coefficients
get_coef1 <- make_coef_getter(mdl) # Using the model object directly
get_coef2 <- make_coef_getter(coefs) # Using the list of coefficients we made

# All are equivalent
get_coef("cyl:gear", coefs)
get_coef1("cyl:gear")
get_coef2("cyl:gear")


}
