% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_helpers.R
\name{get_coef_from_prior}
\alias{get_coef_from_prior}
\title{Extract full coefficient name from the prior}
\usage{
get_coef_from_prior(prior_specs)
}
\arguments{
\item{prior_specs}{List of prior() calls}
}
\value{
Character vector of coefficient names
}
\description{
Extract full coefficient name from the prior
}
