% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets.R
\name{get_reverse_dependencies}
\alias{get_reverse_dependencies}
\title{Get reverse dependencies of a node}
\usage{
get_reverse_dependencies(from = "write_constant_scale_files", edges)
}
\arguments{
\item{from}{Node to search from}

\item{edges}{Edges dataframe from tar_network}
}
\value{
Character vector of target names
}
\description{
Given the name of a target and the edges dataframe from `targets::tar_network()`
return the downstream nodes that depend on the given target. Will recurse
through the entire DAG.
}
