% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helmert_code.R
\name{helmert_code}
\alias{helmert_code}
\title{Helmert code}
\usage{
helmert_code(n_levels)
}
\arguments{
\item{n_levels}{Number of levels in the factor}
}
\value{
Contrast matrix
}
\description{
Since R's contr.helmert implements reverse helmert coding, we can get
regular helmert coding with reversing the reverse helmert matrix.
}
\details{
Helmert coding compares each level to the total mean of all levels that come
after it. Differs from forward difference coding, which only compares pairs
of levels (not a level to a cumulative mean of levels).

Example interpretation for a 4 level factor:
Intercept = Grand mean (mean of the means of each level)
grp1 = mean(grp4, grp3, grp2) - grp(1)
grp2 = mean(grp4, grp3) - mean(grp2)
grp3 = mean(grp3) - mean(grp4)
}
\examples{

mydf <- data.frame(
   grp = factor(c(rep("F1",5),rep("F2",5),rep("F3",5),rep("F4",5))),
   resp = c(seq(1,5), seq(5,9), seq(10,14), seq(15,19))
)

mydf \%>\%
  dplyr::group_by(grp) \%>\%
  dplyr::summarize(mu = mean(resp)) \%>\%
  dplyr::ungroup() \%>\%
  dplyr::mutate(grand_mean = mean(mu))

summary(lm(resp ~ grp,
           data = mydf,
           contrasts = enlist_contrasts(mydf, grp ~ helmert_code)))
}
