% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_centered.R
\name{is_centered}
\alias{is_centered}
\title{Check for orthogonality}
\usage{
is_centered(.contrasts)
}
\arguments{
\item{.contrasts}{Contrast matrix or list of contrast matrices}
}
\value{
Logical vector, will retain names of a passed list
}
\description{
Given a contrast matrix or list of contrast matrices (eg from `enlist_contrasts`),
return a logical vector of whether each contrast is centered or not.
}
