% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendall_tau_b.r
\name{kendall_tau_b}
\alias{kendall_tau_b}
\title{Kendall's tau-b}
\usage{
kendall_tau_b(contingency_table)
}
\arguments{
\item{contingency_table}{2D contingency table of counts}
}
\value{
Kendall's tau b for the sample contingency table.
}
\description{
Calculate sample form of Kendall's Tau-b (Kendall 1945; Agresti 2010, pg 188).
}
\details{
Another association measure based on C-D that also incorporates information
about tied pairs. "In fact, tau-b-hat is a type of correlation even for rxc
tables, using sign socres for pairs of observations." (Agresti 2010, 188)

Note that this is NOT the same as Kendall's tau (Kendall 1938), which should
be used for continuous variables where there are no tied pairs (Tx=Ty=Txy=0).

Because this uses concordancy pairs, ascending or descending order in the
`xtabs` contingency table doesn't matter.
}
