% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coef.R
\name{make_coef_getter}
\alias{make_coef_getter}
\title{Make a coefficient getter}
\usage{
make_coef_getter(obj)
}
\arguments{
\item{obj}{Either a coefficient list from enlist_coefs or a model object}
}
\value{
An estimate string
}
\description{
Create a new function that you can use to retrieve coefficients within
and RMD document.
}
