% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_code.R
\name{manual_code}
\alias{manual_code}
\title{Manual code factors}
\usage{
manual_code(factor_col, coding_matrix = NA)
}
\arguments{
\item{factor_col}{The factor column to use, eg data$gender}

\item{coding_matrix}{The matrix you want to use, obligatory if n factors > 2}
}
\value{
A matrix for contrast coding schemes
}
\description{
Use this when you want to manually specify the contrast matrix
}
