% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_data_grid.R
\name{marginal_data_grid}
\alias{marginal_data_grid}
\title{Create data grid}
\usage{
marginal_data_grid(model, ...)
}
\arguments{
\item{model}{Model fit from `brm`}

\item{...}{Predictor names, optionally set to hold specific levels to predict
with}
}
\value{
Data grid with random effects set at NA and predictors set either
at a specific value or containing all levels
}
\description{
Helper to create a data grid, expands modelr::datagrid
}
\examples{

# Numeric var center_log_freq set to 0,
# Factor var condition uses every level
# Factor var tune uses every level
# So, you'll get every combination of condition and tune held constant at
# centered_log_frequency of 0
\dontrun{
  marginal_data_grid(bmdl, tune, center_log_freq = 0, condition = 'lower')
}
}
