% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spearman_rho_b.R
\name{marginal_ridits}
\alias{marginal_ridits}
\title{Calculate marginal distribution ridits}
\usage{
marginal_ridits(contingency_table)
}
\arguments{
\item{contingency_table}{2D contingency table}
}
\value{
List of marginal ridit values along X and Y
}
\description{
Given a 2D contingency table, calculate the ridit values for the marginal
distributions. This is used to calculate Spearman's rho for rank scored
contingency tables. See Agresti 2010, pg 192.
}
