% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odds_ratios.R
\name{odds_ratios}
\alias{odds_ratios}
\title{Calculate odds ratios from contingency table}
\usage{
odds_ratios(
  table,
  type = "cumulative",
  n = NA,
  use_confint = TRUE,
  bonf = FALSE,
  reverse = FALSE
)
}
\arguments{
\item{table}{Contingency table, recommended to use `xtabs`}

\item{type}{String of which logit to return, either cumulative (default), local, or global}

\item{n}{If using probabilities instead of counts, the total number of observations}

\item{use_confint}{Logical, default TRUE, whether 95% wald confidence intervals
should be calculated and reported in a table}

\item{bonf}{Logical, default FALSE, whether to use Bonferroni correction on standard errors}

\item{reverse}{Logical, default FALSE, whether to reverse the table to be
descending rather than ascending.}
}
\value{
data.frame with odds ratios, standard errors, and confidence intervals
}
\description{
Calculates either local, cumulative, or global odds ratios for a given
contingency table (best made with `xtabs`). Follows the general form equation
given in Agresti, 2010 pg 23. Can also support a table of probabilities rather
than counts so long as `n` is provided.
}
\details{
Note that the interpretation of the resulting `odds_ratio` index changes
depending on the ordering of the factors. `xtabs` will use increasing order
(highest X on the right) while Agresti's textbook uses decreasing order
(highest X on the left). The signs of the estimates are the same regardless
of the order, though. If you would like them to be reversed, use the `reverse`
argument.

95% wald intervals can also be calculated for each logit (Agrsti 2010 pg 27),
and Bonferroni corrections to the standard errors are also supported.
}
