% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomial_code.R
\name{orth_polynomial_code}
\alias{orth_polynomial_code}
\title{Orthogonal Polynomial code}
\usage{
orth_polynomial_code(n_levels)
}
\arguments{
\item{n_levels}{number of levels for this factor}
}
\value{
A matrix of orthogonal polynomial contrasts
}
\description{
Wrapper around R's built in function
}
\details{
For n levels of factors where k in 1:n, generate a matrix with n-1 comparisons where each
comparison looks for a polynomial trend of degree k where each polynomial
is independent of the others.
}
