% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloglog.R
\name{pcloglog}
\alias{pcloglog}
\title{CDF for cloglog distribution}
\usage{
pcloglog(q, location = 0, scale = 1)
}
\arguments{
\item{q}{vector of quantiles}

\item{location}{location parameter}

\item{scale}{scale parameter}
}
\value{
cumulative distribution function
}
\description{
Works like `plogis`, given a vector of numeric quantiles `x` from `(-Inf,Inf)`,
return the cumulative probability up to `x`
}
