% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_helpers.R
\name{posterior_as_new_prior_dists}
\alias{posterior_as_new_prior_dists}
\title{Use posterior parameter estimates as new prior distribution}
\usage{
posterior_as_new_prior_dists(bmdl, scale = 1, shift_to_zero = FALSE)
}
\arguments{
\item{bmdl}{Model fit with brm}

\item{scale}{Numeric multiplier to apply to the standard deviation of the
normal distributions. Set higher than 1 to make priors that are a bit weaker.}

\item{shift_to_zero}{Logical, default FALSE, whether to override the mean for
each prior to be centered at 0 rather than the estimated mean. Useful if you
want regularizing priors with the standard deviation tuned to the posterior
distribution's variance.}
}
\value{
List of prior() calls
}
\description{
Given a brmsfit object, convert each fixed effect parameter into a list of
prior specifications that can be passed to a new model. All parameters are
specified as a normal distribution with mean and standard deviation based
on the draws from all chains of the model.
}
