% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probit_estimates.R
\name{probit_estimates}
\alias{probit_estimates}
\title{Sample probit estimates}
\usage{
probit_estimates(table, reference = 1)
}
\arguments{
\item{table}{Contingency table from `xtabs`}

\item{reference}{Which row should be used as the reference level, defaults to
1}
}
\value{
Data frame with probit thresholds and effect estimates.
}
\description{
Probit analog to `sostools::odds_ratios`. Given a 2d contingency table,
calculate the probit estimates between a reference level (row) and the
other rows. Does not currently support confidence intervals.
}
