% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloglog.R
\name{qcloglog}
\alias{qcloglog}
\title{Inverse CDF for cloglog distribution}
\usage{
qcloglog(p, location = 0, scale = 1)
}
\arguments{
\item{p}{numeric vector of probabilities `(0,1)`}

\item{location}{location parameter}

\item{scale}{scale parameter}
}
\value{
quantile for cloglog distribution of given parameterization
}
\description{
Works like `qlogis`, returns numeric value for cloglog distribution with
given location and scale parameters and a probability `p`
}
