% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomial_code.R
\name{raw_polynomial_code}
\alias{raw_polynomial_code}
\title{Raw Polynomial code}
\usage{
raw_polynomial_code(n_levels)
}
\arguments{
\item{n_levels}{number of levels for this factor}
}
\value{
A matrix of raw polynomial contrasts
}
\description{
Make raw polynomial contrast, rather than orthogonal ones. Normally you
would use orthogonal polynomials, so make sure this is what you want. Using
raw polynomials may increase the collinearity in your model, especially with
higher levels.
}
\details{
For n levels of factors where k in 1:n, generate a matrix with n-1 comparisons where each
comparison looks for a polynomial trend of degree k where each polynomial
may be correlated with the others. Normally you would use orthogonal polynomials.
}
