% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_geom.R
\name{rep_geom}
\alias{rep_geom}
\title{Repeat geom changing a small set of values}
\usage{
rep_geom(geom_call, ...)
}
\arguments{
\item{geom_call}{A geom layer with an aesthetics mapping.}

\item{...}{A named series of values to use for aesthetic mappings.}
}
\value{
list of geoms
}
\description{
Helper to repeat a geom multiple times swapping in different columns or
constant values. Useful for things like drawing a white underdrawing beneath
a layer.
}
\examples{
\dontrun{

rep_geom(geom_line(aes(x = x, y = y),
         linewidth = c(3, 1),
         color = c("white", color_group)))

# == is the same as ==>

list(
  geom_line(aes(x = x, y = y), linewidth = 3, color = "white"),
  geom_line(aes(x = x , y = y, color = color_grp), linewidth = 3)
)


# color is recycled to the length of the other aesthetics
ggplot(mtcars, aes(x = drat, y = mpg)) +
  rep_geom(geom_point(shape = 21),
           size = c(8, 6, 2),
           fill = c("gray80", am, "red"),
           color = "yellow")

# use !! when passing a variable containing a constant value
foo <- "green"
ggplot(mtcars, aes(x = drat, y = mpg)) +
  rep_geom(geom_point(shape = 21),
           size = c(8, 6, 2),
           fill = c(!!foo, am, "red"), # <- just passing foo won't work
           color = "yellow")
           }
}
