% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helmert_code.R
\name{reverse_helmert_code}
\alias{reverse_helmert_code}
\title{Reverse helmert code}
\usage{
reverse_helmert_code(n_levels)
}
\arguments{
\item{n_levels}{Number of levels in the factor}
}
\value{
Contrast matrix
}
\description{
R's contr.helmert function actually implements reverse helmert coding, this
extension scales the matrix such that a 1 unit increase corresponds to a
change in factor level.
}
\details{
Reverse helmert coding compares each level to the total mean of all levels
that have come before it. Differs from backward difference coding, which
compares only pairs of levels (not a level to a cumulative mean of levels)
}
