% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spearman_rho_b.R
\name{rho_b}
\alias{rho_b}
\title{Spearman's rho-b for rank scored contingency tables}
\usage{
rho_b(contingency_table)
}
\arguments{
\item{contingency_table}{2D contingency table}
}
\value{
Sample estimate of spearman's rho-b
}
\description{
"For observations on continuous variables that are ranked on each variable,
Spearman's rho is the ordinary correlation applied to the rank scores.
Kendall (1970, p. 38) proposed an analog of Spearman's rho for contingency
tables. It is the ordinary correlation applied using the ridit scores. [...]
Since the ridit scores are a linear function of the midrakbns, rho-b-hat also
equals the correlation applied to the sample midrank scores. For 2x2 tables,
rho-b = tau-b and then both measures equal the ordinary correlation."
(Agresti 2010, pg 192).
}
