% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_hat.R
\name{sample_gamma}
\alias{sample_gamma}
\title{Calculate sample gamma-hat}
\usage{
sample_gamma(contingency_table)
}
\arguments{
\item{contingency_table}{2D contingency table}
}
\value{
Gamma hat
}
\description{
"Of the C+D pairs of observations that are concordant or discordant, the
proportion of C/(C+D) is concordant and the proportion D/(C+D) is discordant.
THe difference between these proportions is called gamma (Goodman and Kruskal,
1954)" (Agresti 2010, pg 186).
}
\details{
Calculates the sample gamma estimate, which takes values  -1 <= gamma <= 1.
Higher magnitude is stronger association, sign represents positive or negative
association.
}
