% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_contrasts.R
\name{set_contrasts}
\alias{set_contrasts}
\title{Set contrasts to dataframe}
\usage{
set_contrasts(model_data, ...)
}
\arguments{
\item{model_data}{Data to be passed to the model, ensure factors are available}

\item{...}{Series of formulas denoting which contrast scheme to use for each factor}
}
\value{
the model_data dataframe, but with updated contrasts.
}
\description{
Uses the same syntax as enlist_contrasts, but returns the dataframe with the
new contrasts applied. Use this when your model function doesnt have a contrasts
argument and you want to avoid writing contrasts<- multiple times.
}
