% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_groups.R
\name{set_groups}
\alias{set_groups}
\title{Establish plotting groups based on regions of significant differences}
\usage{
set_groups(logicals)
}
\arguments{
\item{logicals}{Logical vector}
}
\value{
Integer indices mapping each value in `logicals` to a group
}
\description{
Given a logical vector, determine groups of contiguous TRUE or FALSE values
}
