% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_helpers.R
\name{show_posterior_approximations}
\alias{show_posterior_approximations}
\title{Plot normal approximations of the posterior distribution}
\usage{
show_posterior_approximations(
  bmdl,
  params = "",
  scale = 1,
  shift_to_zero = FALSE
)
}
\arguments{
\item{bmdl}{brmsfit object}

\item{params}{Which parameters to restrict the plot to}

\item{scale}{Numeric multiplier for the standard deviations of parameters}

\item{shift_to_zero}{Whether to shift the means of the posterior normal
approximations to 0}
}
\value{
List of geoms to add to the hdi plot
}
\description{
This is a helper to plot normal distributions on top of of the plot created
by `plot(bayestestR::hdi(model))`. Useful when you're using the posterior
distribution of some parameter as the prior for another model, but with the
standard deviation scaled or the mean shifted to zero.
}
\examples{
\dontrun{

plot(bayestestR::hdi(model)) +
  show_posterior_approximations(model)
}

}
