% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somers_d.R
\name{somers_d}
\alias{somers_d}
\title{Calculate Somers' d}
\usage{
somers_d(contingency_table)
}
\arguments{
\item{contingency_table}{2D contingency table}
}
\value{
Sample estimate of Somers' d
}
\description{
"Somers (1962) proposed a measure ismilar to gamma and tau-b, but which treats
Y as a response variable and X as an explanatory variable." (Agresti 2010,
189).
}
