% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/png_watermarking.R
\name{string_to_binary}
\alias{string_to_binary}
\title{Convert a string to binary character vector}
\usage{
string_to_binary(input_string)
}
\arguments{
\item{input_string}{String to encode}
}
\value{
Named character vector of binary encodings
}
\description{
Given a string, convert each character to binary representation. Limited to
8 bit representations, which is more than enough for ascii characters.
}
