% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_code.R
\name{sum_code}
\alias{sum_code}
\title{Sum code}
\usage{
sum_code(n_levels)
}
\arguments{
\item{n_levels}{number of levels for this factor}
}
\value{
A matrix of sum coded contrasts, unlabeled
}
\description{
For n levels of factors, generate a matrix with n-1 comparisons where:
Reference level = -1
Comparison level = 1
All others = 0
}
