% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_as_tex.R
\name{summary_as_tex}
\alias{summary_as_tex}
\title{Model Summary as table}
\usage{
summary_as_tex(
  model,
  correct = NA,
  statistic = "$z$",
  caption = NA,
  label = NA
)
}
\arguments{
\item{model}{Model object}

\item{correct}{Regular expressions targeting coefficients to bonferroni correct}

\item{statistic}{Header label for the test statistic, recommended to bookend
in $s}

\item{caption}{Caption to use for the table}

\item{label}{Label to use when referencing the table}
}
\description{
This is just a quick way to make a table for output to either html or, more
importantly, latex fragments. If you want more fine grained customization,
use gtsummary or kable yourself. Right now these have to be glmers or lmers,
but I'll make something more general later when/if I need it.
}
