% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets.R
\name{tar_load_em}
\alias{tar_load_em}
\title{Load model from error message}
\usage{
tar_load_em()
}
\value{
nothing, loads model to current environment
}
\description{
Same as tar_load_err() but with models specifically
}
