% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets.R
\name{tar_load_model}
\alias{tar_load_model}
\title{Load model file to current environment}
\usage{
tar_load_model(model_target_name)
}
\arguments{
\item{model_target_name}{Name of the target containing a path to an rds file}
}
\value{
nothing, assigns to current environment
}
\description{
Given a model name, load the model to the current environment.
eg: tar_load_model(exp1_pabt_WKSLP_mdl)
}
