% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets.R
\name{tar_read_model}
\alias{tar_read_model}
\title{Read model}
\usage{
tar_read_model(model_target_name)
}
\arguments{
\item{model_target_name}{}
}
\value{
object saved at the rds file
}
\description{
Given a model name (=target name), return the model object associated with it.
Note that the stored value of the TARGET is the PATH to the model, so it
needs to be read with readRDS. This just wraps around that 2-step process.
}
