% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_propodds.R
\name{test_propodds}
\alias{test_propodds}
\title{Proportional odds Testing}
\usage{
test_propodds(ord_data, model_formula)
}
\arguments{
\item{ord_data}{Data frame with your data}

\item{model_formula}{The model formula to provide to a glm call}
}
\value{
A data frame with estimates and confidence intervals for each scale response
}
\description{
Tests for proportional odds assumption in data. You can use as many predictors
as you want, but you cannot use random effects (only fixed effects allowed)
for the time being. A message will be displayed whether if the proportional odds
assumption is likely met or not. A dataframe with estimates and confidence intervals
will be returned if you want to plot th rsults to double check.
}
