% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendall_tau_b.r
\name{tied_pairs}
\alias{tied_pairs}
\title{Number of tied pairs}
\usage{
tied_pairs(contingency_table)
}
\arguments{
\item{contingency_table}{2D Contingency table}
}
\value{
Named vector of the number of tied pairs Tx and Ty
}
\description{
Calculate the number of tied pairs in a contingency table, see Agresti 2010
pg 188. Txy is not calculated because it factors out of other calculations.
This is used in the calculation of Kendall's tau-b
}
