% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatment_code.R
\name{treatment_code}
\alias{treatment_code}
\title{Treatment code}
\usage{
treatment_code(n_levels)
}
\arguments{
\item{n_levels}{number of levels for this factor}
}
\value{
A matrix of treatment coded contrasts, unlabeled
}
\description{
Wrapper around R's built in function
}
\details{
For n levels of factors, generate a matrix with n-1 comparisons where:
Reference level = 0
Comparison level = 1
}
