% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_helpers.R
\name{unpack_prior_specification}
\alias{unpack_prior_specification}
\title{Convert prior specification to dataframe}
\usage{
unpack_prior_specification(prior_specs)
}
\arguments{
\item{prior_specs}{List of prior()}
}
\value{
Dataframe with prior information
}
\description{
Given a list of priors, convert them to a dataframe.
brms represents priors as dataframes anyways, but this adds a few additional
columns of information. Namely, the full coefficient name, the distribution
used, and the arguments to that distribution as a list column.
}
