% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/png_watermarking.R
\name{watermark_png}
\alias{watermark_png}
\title{Watermark a png}
\usage{
watermark_png(from, to = NA)
}
\arguments{
\item{from}{Path to png file}

\item{to}{Output path file, set to NA to overwrite file specified by `from`}
}
\value{
output filepath
}
\description{
Reads in a png file and watermarks the top left corner with the current
system time. Watermark is 26x8 pixels large.
}
