% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/png_watermarking.R
\name{watermark_png_figure}
\alias{watermark_png_figure}
\title{Watermark a png figure}
\usage{
watermark_png_figure(files, strict = FALSE)
}
\arguments{
\item{files}{Character vector containing an element named png}

\item{strict}{Logical, set to TRUE to throw an error if no element named
png is found, otherwise a warning is given.}
}
\value{
files
}
\description{
Given a named character vector with an element named 'png', watermark the
file and return the path. Intended to serve in a pipeline after `save_plot`
}
