% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward_difference_code.R
\name{backward_difference_code}
\alias{backward_difference_code}
\title{Backward difference code}
\usage{
backward_difference_code(n_levels)
}
\arguments{
\item{n_levels}{Number of factor levels}
}
\value{
Backward difference contrast matrix
}
\description{
Compares the mean of level k to level k-1. Differs from reverse helmert coding
because it doesn't compare k to mean(k_0, ..., k_i-1). Mathematically just
flipping the signs of the matrix from forward difference coding.
}
\details{
Intercept is the grand mean (mean of cell means)
}
