% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast_code.R
\name{contrast_code}
\alias{contrast_code}
\title{Contrast Code with Labels}
\usage{
contrast_code(factor_col, coding_matrix = NA)
}
\arguments{
\item{factor_col}{The factor column to use, eg data$gender}

\item{coding_matrix}{The matrix you want to use, obligatory if n factors > 2}
}
\value{
A matrix for contrast coding schemes
}
\description{
This makes it easier to add contrast coding schemes while retaining the
labels of the levels IN THE ORDER THEY APPEAR IN THE ORIGINAL CONTRAST
SCHEME. Usually in alphabetical order. Will default to scaled sum coding
for 2 level factors with first alphabetically as the reference level. If
it's greater than 2 levels you MUST provide your own matrix using the matrix
function.
}
