% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glimpse_contrasts.R
\name{.glimpse_default_factors}
\alias{.glimpse_default_factors}
\title{Glimpse default factors}
\usage{
.glimpse_default_factors(model_data, set_factors)
}
\arguments{
\item{model_data}{Dataframe}

\item{set_factors}{Explicitly set columns to ignore}
}
\value{
A table with information about the contrasts for all remaining factor
columns
}
\description{
Given a dataframe with some factor columns and a character vector of
which columns you've already set yourself, look at all the other
factor columns and get a glimpse at how they're treated by the defaults
specified in `options('contrasts')`. Reference level is assumed to be the
first level for unordered factors and nonexistent for ordered factors.
}
