% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_formula.R
\name{.parse_formula}
\alias{.parse_formula}
\title{Parse contrast formula}
\usage{
.parse_formula(raw_formula, has_matrix = FALSE)
}
\arguments{
\item{raw_formula}{Raw formula passed by user}

\item{has_matrix}{Whether a matrix call is detected or not}
}
\value{
A list of parameters to use for a contrast_code call
}
\description{
Takes a formula and figures out the parameters to use based on the operands
provided in the formula
}
