% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enlist_contrasts.R
\name{.process_contrasts}
\alias{.process_contrasts}
\title{Pass arguments to contrast code}
\usage{
.process_contrasts(model_data, char_formula, var_envir)
}
\arguments{
\item{model_data}{Data frame with factor column}

\item{char_formula}{character conversion of formula}

\item{var_envir}{Environment in which the reference variables can be found}
}
\value{
A contrast matrix
}
\description{
Helper to generate contrast matrix from passed formulas in enlist_contrasts
}
