% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enlist_coefs.R
\name{enlist_coefs}
\alias{enlist_coefs}
\title{Enlist Coefficients}
\usage{
enlist_coefs(model, correct = NA, ...)
}
\arguments{
\item{model}{A model object with a summary method}

\item{correct}{Regular expressions targeting coefficients to bonferroni correct}

\item{...}{Additional arguments if model is a brmsfit, see docs for tidy_brmsfit for options}
}
\value{
A list of coefficient descriptions of the form (Beta = b, s.e.=se, z=z, p=p)
for latex output
}
\description{
Takes a model object and creates a list of coefficients that can be referenced
throughout an R markdown file.
}
\examples{
enlist_coefs(lm(mpg~cyl, data = mtcars))
}
