% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward_difference_code.R
\name{forward_difference_code}
\alias{forward_difference_code}
\title{Forward difference code}
\usage{
forward_difference_code(n_levels)
}
\arguments{
\item{n_levels}{Number of factor levels}
}
\value{
Forward difference contrast matrix
}
\description{
Compares the mean of level k to level k+1. Differs from helmert coding because
it doesn't compare k to mean(k_i, ..., k_n)
Intercept is the grand mean (mean of cell means)
}
