% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helmert_code.R
\name{helmert_code}
\alias{helmert_code}
\title{Helmert code}
\usage{
helmert_code(n_levels)
}
\arguments{
\item{n_levels}{Number of levels in the factor}
}
\value{
Contrast matrix
}
\description{
Since R's contr.helmert implements reverse helmert coding, we can get
regular helmert coding with reversing the reverse helmert matrix.
}
\details{
Helmert coding compares each level to the total mean of all levels that come
after it. Differs from forward difference coding, which only compares pairs
of levels (not a level to a cumulative mean of levels).
}
