% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_propodds.R
\name{plot_propodds}
\alias{plot_propodds}
\title{Plot proportional odds output}
\usage{
plot_propodds(propodds_results)
}
\arguments{
\item{propodds_results}{Output of test_propodds}
}
\description{
A very simple helper to plot the estimates from the proportional odds tester.
}
