% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helmert_code.R
\name{reverse_helmert_code}
\alias{reverse_helmert_code}
\title{Reverse helmert code}
\usage{
reverse_helmert_code(n_levels)
}
\arguments{
\item{n_levels}{Number of levels in the factor}
}
\value{
Contrast matrix
}
\description{
R's contr.helmert function actually implements reverse helmert coding, this
extension scales the matrix such that a 1 unit increase corresponds to a
change in factor level.
}
\details{
Reverse helmert coding compares each level to the total mean of all levels
that have come before it. Differs from backward difference coding, which
compares only pairs of levels (not a level to a cumulative mean of levels)

Example interpretation for a 4 level factor:
\itemize{
\item Intercept = Grand mean (mean of the means of each level)
\item grp2 = mean(grp2) - mean(grp1)
\item grp3 = mean(grp3) - mean(grp1, grp2)
\item grp4 = mean(grp4) - mean(grp1, grp2, grp3)
}
}
\examples{

mydf <- data.frame(
   grp = factor(c(rep("F1",5),rep("F2",5),rep("F3",5),rep("F4",5))),
   resp = c(seq(1,5), seq(5,9), seq(10,14), seq(15,19))
)

mydf \%>\%
  dplyr::group_by(grp) \%>\%
  dplyr::summarize(mu = mean(resp)) \%>\%
  dplyr::ungroup() \%>\%
  dplyr::mutate(grand_mean = mean(mu))

summary(lm(resp ~ grp,
           data = mydf,
           contrasts = enlist_contrasts(mydf, grp ~ reverse_helmert_code)))
}
