% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaled_sum_code.R
\name{scaled_sum_code}
\alias{scaled_sum_code}
\title{Scaled sum coding}
\usage{
scaled_sum_code(n_levels)
}
\arguments{
\item{n_levels}{number of levels for this factor}
}
\value{
A matrix of scaled sum coded contrasts, unlabeled
}
\description{
For n levels of factors, generate a matrix with n-1 comparisons where:
Reference level = -1/n
Comparison level = (n-1)/n
All others = -1/n
}
