% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_as_tex.R
\name{summary_as_tex}
\alias{summary_as_tex}
\title{Model Summary as table}
\usage{
summary_as_tex(
  model,
  correct = NA,
  statistic = "$z$",
  caption = NA,
  label = NA,
  table.envir = "table",
  tabular.envir = "tabular",
  include_prior = FALSE,
  ci_level = 0.95,
  use_ROPE = FALSE
)
}
\arguments{
\item{model}{Model object}

\item{correct}{Regular expressions targeting coefficients to bonferroni correct}

\item{statistic}{Header label for the test statistic, recommended to bookend
in $s}

\item{caption}{Caption to use for the table}

\item{label}{Label to use when referencing the table}

\item{table.envir}{LaTeX table environment to use, defaults to table, change
not implemented yet}

\item{tabular.envir}{LaTeX tabular environment to use, defaults to table, can
be changed to tabular*}

\item{include_prior}{Should priors be included? (only if model is brmsfit)}

\item{ci_level}{Defaults to .95 (only if model is brmsfit)}

\item{use_ROPE}{ROPE object from bayestestR::rope, or TRUE to run this with the
default settings. Adds the % in ROPE to table after confidence intervals. FALSE
to not include. (only if model is brmsfit)}
}
\description{
This is just a quick way to make a table for output to either html or, more
importantly, latex fragments. If you want more fine grained customization,
use gtsummary or kable yourself. Right now these have to be glmers or lmers,
but I'll make something more general later when/if I need it.
}
